Faster R-CNN/requirements.txt
visdom
torchvision
scikit-image
tqdm
fire
matplotlib
ipdb
cython
git+https://github.com/pytorch/tnt.git@master